refactor: put all view components in View folder
