--- src/components/board/BoardDetail.spec.ts
@@ -1,5 +1,0 @@
-import BoardDetail from '@/components/board/BoardDetail.vue';
-
-it('shows BoardDetail', () => {
-  cy.mount(BoardDetail);
-});
