+++ src/components/board/BoardHeader.vue
@@ -1,35 +1,32 @@
 <template>
-      <div class="py-2.5">
-        <div class="inline-block relative py-1.5 mr-0 ml-3 h-8">
-          <div class="inline-block invisible px-3 font-bold">
-            <!-- {{ state.board.name }} -->
-            {{ boardName }}
-          </div>
-
-          <!-- v-click-away="onClickAway" -->
-          <input
-            v-model="boardName"
-            class="text-white bg-white bg-opacity-20 hover:bg-opacity-30 board-title"
-            data-cy="board-title"
-            autocomplete="off"
-            name="board-title"
-            @focus="selectInput($event)"
-            @change="onChangeBoardName"
-            @keyup.enter="blurInput($event)"
-            @keyup.esc="blurInput($event)"
-          >
-
-        </div>
-        <div
-          class="inline-grid relative self-center ml-2 w-8 h-8 bg-white bg-opacity-20 hover:bg-opacity-30 rounded-sm cursor-pointer"
-          :class="[starred ? 'fill-current text-yellow-300' : 'stroke-current text-white']"
-          data-cy="star"
-          @click="$emit('clickOnStar')"
-        >
-          <Star class="place-self-center m-2" />
-        </div>
-        <BoardOptions :board="board" />
+  <div class="py-2.5">
+    <div class="inline-block relative py-1.5 mr-0 ml-3 h-8">
+      <div class="inline-block invisible px-3 font-bold">
+        {{ boardName }}
       </div>
+
+      <input
+        v-model="inputBoardName"
+        class="text-white bg-white bg-opacity-20 hover:bg-opacity-30 board-title"
+        data-cy="board-title"
+        autocomplete="off"
+        name="board-title"
+        @focus="selectInput($event)"
+        @change="onChangeBoardName"
+        @keyup.enter="blurInput($event)"
+        @keyup.esc="blurInput($event)"
+      >
+    </div>
+    <div
+      class="inline-grid relative self-center ml-2 w-8 h-8 bg-white bg-opacity-20 hover:bg-opacity-30 rounded-sm cursor-pointer"
+      :class="[starred ? 'fill-current text-yellow-300' : 'stroke-current text-white']"
+      data-cy="star"
+      @click="$emit('clickOnStar')"
+    >
+      <Star class="place-self-center m-2" />
+    </div>
+    <BoardOptions :board="board" />
+  </div>
 </template>
 
 <script setup lang="ts">
@@ -38,17 +35,21 @@ import { blurInput } from '@/utils/blurInput';
 import { selectInput } from '@/utils/selectInput';
 import BoardOptions from '@/components/board/BoardOptions.vue';
 import Star from '@/assets/icons/star.svg';
+import { ref, watch } from 'vue';
 
 const props = defineProps<{
-    boardName: string,
-    board: Board,
-    starred: boolean,
-    onChangeBoardName: () => void
-}>()
+  boardName: string;
+  board: Board;
+  starred: boolean;
+  onChangeBoardName: () => void;
+}>();
 
-console.log('starred', props.starred);
+const emit = defineEmits(['update:boardName', 'clickOnStar']);
 
-defineEmits(['update:boardName', 'clickOnStar']);
+const inputBoardName = ref(props.boardName);
+watch(inputBoardName, () => {
+  emit('update:boardName', inputBoardName.value);
+})
 </script>
 
 <style lang="postcss" scoped>
