+++ src/components/board/BoardLoading.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="grid justify-center content-center h-screen loading">
-      <div><LoadingIcon class="inline-block mb-1" />&nbsp;&nbsp;Loading data ...</div>
-    </div>
+  <div class="grid justify-center content-center h-screen loading">
+    <div><LoadingIcon class="inline-block mb-1" />&nbsp;&nbsp;Loading data ...</div>
+  </div>
 </template>
