+++ src/views/BoardList.vue
